// ui/src/api/gen/SidebarPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LocalBranch } from "./LocalBranch";
import type { RemoteGroup } from "./RemoteGroup";
import type { StashItem } from "./StashItem";
import type { TagItem } from "./TagItem";
import type { WorktreeItem } from "./WorktreeItem";

export type SidebarPayload = { locals: Array<LocalBranch>, remotes: Array<RemoteGroup>, worktrees: Array<WorktreeItem>, stashes: Array<StashItem>, tags: Array<TagItem>, };